back/aba-admissions-manager/requirements.txt
numpy==1.25.1
pandas==2.0.3
psycopg2_binary==2.9.7
PyYAML==6.0.1
scikit_learn==1.1.3
snowflake_connector_python[secure-local-storage,pandas]==3.0.4
xgboost==2.0.0

